+++ riscv/examples/riscv.rs
@@ -8,6 +8,7 @@ use std::{
     collections::{BTreeMap, HashMap},
     fs,
     path::PathBuf,
+    process::Command,
     rc::Rc,
 };
 use syncrim::{
@@ -18,9 +19,6 @@ use xmas_elf::ElfFile;
 
 #[derive(Parser, Debug)]
 struct Args {
-    /// Toolchain prefix to be used for compilation/linking
-    #[arg(short, long, default_value = "riscv32-unknown-elf-")]
-    toolchain_prefix: String,
     /// Use a pre-compiled elf file instead of compiling one
     #[arg(short, long, default_value = "false")]
     use_elf: bool,
@@ -38,17 +36,18 @@ struct Args {
 fn main() {
     fern_setup_riscv();
     let args = Args::parse();
-    let memory = if args.use_elf {
-        let bytes = fs::read(args.elf_path).expect("The elf file could not be found");
+    let memory = if !args.use_elf {
+        elf_from_asm(&args);
+        let bytes = fs::read("./output").expect("The elf file could not be found");
         let elf = ElfFile::new(&bytes).unwrap();
         riscv_elf_parse::Memory::new_from_elf(elf)
     } else {
-        riscv_elf_parse::Memory::new_from_assembly(
-            &args.asm_path,
-            &args.ls_path,
-            &args.toolchain_prefix,
-        )
+        let bytes =
+            fs::read(format!("{}", args.elf_path)).expect("The elf file could not be found");
+        let elf = ElfFile::new(&bytes).unwrap();
+        riscv_elf_parse::Memory::new_from_elf(elf)
     };
+
     println!("{}", memory);
     let mut instr_mem = BTreeMap::new();
     let mut data_mem = HashMap::new();
@@ -140,86 +139,6 @@ fn main() {
                 pc: Input::new("reg", "out"),
                 // fake instructions just to show the relation between input address and instruction
                 bytes: instr_mem,
-                //vec![
-                //The results are calculated from the reg file start state defined above
-                //and assuming the blocks are the only code over that reg file.
-                //OP TEST BLOCK
-                // 0x003100b3,//add x1, x2, x3 -> x1 = 0x5
-                // 0x00308133,//add x2, x1, x3 -> x2 = 0x8
-                // 0x40410133,//sub x2, x2, x4 -> x2 = -2
-                // 0x004121b3,//slt x3, x2, x4 -> x3 = 0x1 //signed -2<10
-                // 0x004131b3,//sltu x3, x2, x4 x3 = 0x0 //unsigned -2>10
-                // 0x001151b3,//srl x3, x2, x1 # x3 = 0x07ffffff
-                // 0x401151b3,//sra x3, x2, x1 # x3 = 0xffffffff
-                // 0x001111b3,//sll x3, x2, x1 # x3 = 0xffffffc0
-                // 0x0020c1b3,//xor x3, x1, x2 # x3 = 0xfffffffb //fel här
-                // 0x0020f1b3,//and x3, x1, x2 # x3 = 0x00000004
-                // 0x0060e1b3,//or x3, x1, x6 #  x3 = 0x00000007
-                // 0x00000033,//add x0, x0, x0, basically nop before panicking so we can see result.
-                // 0x00940023,//sb x8, 0(x9) # should panic over opcode for now
-                //OP_IMM, AUIPC, LUI, STORE, LOAD, OP_IMM TEST BLOCK
-                // 0x00310093,//addi x1, x2, 3 # x1=0x5
-                // 0xffd0a093,//slti x1, x1, -3 # x1=0x0
-                // 0x0030a093,//slti x1, x1, 3 # x1=0x1
-                // 0xffd0b093,//sltiu x1, x1, -3 #x1=0x1
-                // 0x00313093,//sltiu x1, x2, 3 #x1=0x1
-                // 0x00324093,//xori x1, x4, 3 #x1 = 0x9
-                // 0x00326093,//ori x1, x4, 3 #x1=0xb
-                // 0x00327093,//andi x1, x4, 3 #x1=0x2
-                // 0x00c19093,//slli x1, x3, 12 #x1=0x3000
-                // 0x0011d093,//srli x1, x3, 1 #x1=0x1
-                // 0xffa00093,//addi x1, x0, -6 #x1=0xfffffffa
-                // 0x4020d093,//srai x1, x1, 2 #x1=0xfffffffe
-                // 0x00500093,//addi x1, x0, 5 #x1=0x5
-                // 0x4020d093,//srai x1, x1, 2 #x1=0x1
-                // 0xfffff0b7,//lui x1, 0xFFFFF #x1=0xFFFFF000
-                // 0xfffff097,//auipc x1, 0xFFFFF #x1=0xFFFFF040
-                // 0x00000093,//addi, x1, x0, 0 x1 = 0
-                // 0x00408093,//addi x1, x1, 4 x1+=4
-                // //0xff9ff16f,//jal, x2, -8 should jump to the addi before and keep incrementing x1.
-                // 0x00000033,//add x0, x0, x0, basically nop before panicking so we can see result.
-                // //0x00940023,//sb x8, 0(x9) # should panic over opcode for now
-                // 0x00102023,//sw x1, 0(x0) store x1=4 at 0
-                // 0x00002283,//lw x5, 0(x0) x5=4
-                // 0x00228213,//addi x4, x5, 2 //x4=6
-                // 0x00002303, //lw x6, 0(x0), x6 = 4
-                // 0xfff00093,//set x1 to -1 addi x1, x0, -1
-                // 0x00102023,//store -1 at 0
-                // 0x00000203,//load via lb -1 to x4
-                // 0x00004203,//lbu x4 -1 again.
-                // 0x004002a3,//sb x4, 5(0)
-                // 0x00000033,//add x0, x0, x0, basically nop before panicking so we can see result.
-                // 0x00000033,//add x0, x0, x0, basically nop before panicking so we can see result.
-                //JAL, JALR, BRANCHES TEST BLOCK
-                //     0x00000093, //addi, x1, x0, 0 x1=0
-                //     0x0080016f, //jal x2, 8
-                //     0x0000016f, //jal x2, 0 this is to be jumped over or we will get stuck
-                //     0x00410167, //jalr x2, x2, 4
-                //     0x0000016f, //jal x2, 0 this is to be jumped over or we will get stuck
-                //     0x01000113, //addi x2, x0, 16
-                //     0x00000093, //addi x1, x0, 0
-                //     0x00408093, //addi x1, x1, 4
-                //     0xfe209ee3, //bne x1, x2, -4
-                //     0x00208463, //beq, x1, x2, 8
-                //     0x0000006f, //jal x0, 0
-                //     0xff800093, //addi x1, x0, -8
-                //     0x00000113, //addi x2, x0, 0
-                //     0x00408093, //addi x1, x1, 4
-                //     0xfe20cee3, //blt x1, x2, -4
-                //     0xff800093, //addi x1, x0, -8
-                //     0x00000113, //addi x2, x0, 0
-                //     0x00408093, //addi x1, x1, 4
-                //     0xfe116ee3, //bltu, x2, x1, -4
-                //     0xff800093, //addi x1, x0, -8
-                //     0x00000113, //addi x2, x0, 0
-                //     0x00408093, //addi x1, x1, 4
-                //     0xfe115ee3, //bge x2, x1, -4
-                //     0xff800093, //addi x1, x0, -8
-                //     0x00800113, //addi x2, x0, 8
-                //     0x00408093, //addi x1, x1, 4
-                //     0xfe20fee3, //bgeu x1, x2, -4
-                //     0x00408093, 0x00408093, 0x00408093, 0x00408093, 4, 5, 6, 7, 8, 9,
-                // ],
             }),
             Rc::new(Decoder {
                 id: "decoder".to_string(),
@@ -252,43 +171,7 @@ fn main() {
                 write_data: Input::new("wb_mux", "out"),
                 write_addr: Input::new("regfile_rd_reg", "out"),
                 write_enable: Input::new("regfile_we_reg", "out"),
-                registers: RegStore::new(Rc::new(RefCell::new([
-                    0, 1, 2, 3, 10, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
-                    22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
-                ]))),
-                // registers: vec![Cell::new(0),
-                // RefCell::new(1),
-                // Cell::new(2),
-                // Cell::new(3),
-                // Cell::new(10),
-                // Cell::new(5),
-                // Cell::new(6),
-                // Cell::new(7),
-                // Cell::new(8),
-                // Cell::new(9),
-                // Cell::new(10),
-                // Cell::new(11),
-                // Cell::new(12),
-                // Cell::new(13),
-                // Cell::new(14),
-                // Cell::new(15),
-                // Cell::new(16),
-                // Cell::new(17),
-                // Cell::new(18),
-                // Cell::new(19),
-                // Cell::new(20),
-                // Cell::new(21),
-                // Cell::new(22),
-                // Cell::new(23),
-                // Cell::new(24),
-                // Cell::new(25),
-                // Cell::new(26),
-                // Cell::new(27),
-                // Cell::new(28),
-                // Cell::new(29),
-                // Cell::new(30),
-                // Cell::new(31),
-                // ],
+                registers: RegStore::new(Rc::new(RefCell::new([0; 32]))),
                 history: RegHistory::new(),
             }),
             Mem::rc_new_from_bytes(
@@ -323,6 +206,7 @@ fn main() {
                     Input::new("reg_file", "reg_b"),
                     Input::new("imm_szext", "out"),
                     Input::new("pc_adder", "out"),
+                    Input::new("reg", "out"),
                 ],
             ),
             Rc::new(ALU {
@@ -384,4 +268,73 @@ fn fern_setup_riscv() {
         // Apply globally
         .apply()
         .unwrap()
+}
+
+fn elf_from_asm(args: &Args) {
+    let source_path = &args.asm_path;
+    let linker_path = &args.ls_path;
+    let _ = if cfg!(target_os = "windows") {
+        Command::new("cmd")
+            .current_dir(".\\")
+            .args(["/C", &format!("copy {} .\\riscv_asm\\asm.s", source_path)])
+            .status()
+            .unwrap()
+    } else {
+        Command::new("sh")
+            .current_dir("./")
+            .arg("-c")
+            .arg(format!("cp {} ./riscv_asm/asm.s", source_path))
+            .status()
+            .unwrap()
+    };
+    let _ = if cfg!(target_os = "windows") {
+        Command::new("cmd")
+            .current_dir(".\\")
+            .args([
+                "/C",
+                &format!("copy {} .\\riscv_asm\\memory.x", linker_path),
+            ])
+            .status()
+            .unwrap()
+    } else {
+        Command::new("sh")
+            .current_dir("./")
+            .arg("-c")
+            .arg(format!("cp {} ./riscv_asm/memory.x", linker_path))
+            .status()
+            .unwrap()
+    };
+    let _ = if cfg!(target_os = "windows") {
+        Command::new("cmd")
+            .current_dir(".\\riscv_asm\\")
+            .args(["/C", "cargo build --release"])
+            .status()
+            .unwrap()
+    } else {
+        Command::new("sh")
+            .current_dir("./riscv_asm/")
+            .arg("-c")
+            .arg(format!("cargo build --release"))
+            .status()
+            .unwrap()
+    };
+    let _ = if cfg!(target_os = "windows") {
+        Command::new("cmd")
+            .current_dir(".\\riscv_asm\\")
+            .args([
+                "/C",
+                "move /y .\\target\\riscv32i-unknown-none-elf\\release\\riscv_asm ..\\output",
+            ])
+            .status()
+            .unwrap()
+    } else {
+        Command::new("sh")
+            .current_dir("./")
+            .arg("-c")
+            .arg(format!(
+                "mv ./riscv_asm/target/riscv32i-unknown-none-elf/release/riscv_asm ./output"
+            ))
+            .status()
+            .unwrap()
+    };
 }